Uma quase vitória

+++ app/Http/Controllers/CategoryController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Category;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CategoryController extends Controller
 {
@@ -23,11 +25,13 @@ class CategoryController extends Controller
 
     public function create()
     {
+        Gate::authorize('bibliotecario', User::class);
         return view('categories.create');
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('bibliotecario', User::class);
         $validatedData = $request->validate([
             'name' => 'required|string|max:255|unique:categories',
         ]);
@@ -40,6 +44,7 @@ class CategoryController extends Controller
 
     public function edit($id)
     {
+        Gate::authorize('bibliotecario', User::class);
         $category = Category::findOrFail($id);
         return view('categories.edit', compact('category'));
 
@@ -50,6 +55,7 @@ class CategoryController extends Controller
      */
     public function update(Request $request, $id)
     {
+        Gate::authorize('bibliotecario', User::class);
        $validatedData = $request->validate([
             'name' => 'required|string|max:255|unique:categories,name,' . $id,
                ]);
@@ -66,6 +72,7 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('bibliotecario', User::class);
        $category = Category::findOrFail($id);
                $category->books()->detach();
                $category->delete();
